wip: separated pages logic
